Move base functions to maintenance module to import other places
